Move Marks of Chaos to alliances file
