+++ src/army/slaves_to_darkness/traits.ts
@@ -9,7 +9,6 @@ const CommandTraits: TTraits = [
         name: `Eternal Vendetta`,
         desc: `You can re-roll wound rolls for attacks made with melee weapons by this general that target an Order unit.`,
         when: [COMBAT_PHASE],
-        command_trait: true,
       },
     ],
   },
@@ -20,7 +19,6 @@ const CommandTraits: TTraits = [
         name: `Flames of Spite`,
         desc: `If the unmodified wound roll for an attack made by this general is 6, the target suffers 1 mortal wound in addition to any normal damage.`,
         when: [SHOOTING_PHASE, COMBAT_PHASE],
-        command_trait: true,
       },
     ],
   },
@@ -31,7 +29,6 @@ const CommandTraits: TTraits = [
         name: `Master of Deception`,
         desc: `Subtract 1 from hit rolls for attacks made with melee weapons that target this general.`,
         when: [COMBAT_PHASE],
-        command_trait: true,
       },
     ],
   },
@@ -42,7 +39,6 @@ const CommandTraits: TTraits = [
         name: `Hatred Incarnate`,
         desc: `You can re-roll wound rolls of 1 for attacks made by this general.`,
         when: [SHOOTING_PHASE, COMBAT_PHASE],
-        command_trait: true,
       },
     ],
   },
@@ -53,7 +49,6 @@ const CommandTraits: TTraits = [
         name: `Lord of Terror`,
         desc: `Subtract 1 from the bravery characterisic of enemy units while they are within 6" of this general.`,
         when: [DURING_GAME],
-        command_trait: true,
       },
     ],
   },
@@ -64,7 +59,6 @@ const CommandTraits: TTraits = [
         name: `Exalted Champion`,
         desc: `Add 1 to the bravery characterisic of friendly Slaves to Darkness units while they are wholly within 12" of this general.`,
         when: [DURING_GAME],
-        command_trait: true,
       },
     ],
   },
